requirements.txt
astroid==2.6.6
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Deprecated==1.2.12
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
pandas==1.3.2
plotly==5.2.1
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
pylint==2.9.6
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
tenacity==8.0.1
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
